refactor: remove Theme barrel export and use direct imports
