--- src/components/Theme/index.tsx
@@ -1,2 +1,0 @@
-export { default as Provider } from "@/components/Theme/Provider";
-export { default as Toggle } from "@/components/Theme/Toggle";
